UPSTREAM: mm/page_alloc.c: fix the value of fallback_migratetype in alloc_extfrag tracepoint()
(cherry pick from commit f92310c1877fc73470bdcd9228758fa3713c191b)

In the current code, the value of fallback_migratetype that is printed
using the mm_page_alloc_extfrag tracepoint, is the value of the
migratetype *after* it has been set to the preferred migratetype (if the
ownership was changed).  Obviously that wouldn't have been the original
intent.  (We already have a separate 'change_ownership' field to tell
whether the ownership of the pageblock was changed from the
fallback_migratetype to the preferred type.)

The intent of the fallback_migratetype field is to show the migratetype
from which we borrowed pages in order to satisfy the allocation request.
So fix the code to print that value correctly.

Signed-off-by: Srivatsa S. Bhat <srivatsa.bhat@linux.vnet.ibm.com>
Cc: Mel Gorman <mel@csn.ul.ie>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Cody P Schafer <cody@linux.vnet.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Bug: 26295735
Bug: 26412925
Bug: 26379045
Change-Id: I3528d8cf15f9d9152afd9818b79e58433b833e29
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>

(cherry picked from commit 2a42efabafdf41200593224ed4488512c271e804)