Update Barry Allen 0.9 Fix for revert (add calculate load before freq change)
The update to cpu load for cpufreq rate changes was happening after the rate
change instead of before. This switches it to update the cpu load before the
cpufreq rate change so the old frequency is used for calculating the speed
adjusted cpu load. The old frequency was used for that active time, so it should
be used for calculating the speed adjusted cpu load.

Signed-off-by: Javier Sayago <admin@lonasdigital.com>
Signed-off-by: AudioGod <audiogod@sonic-developers.com>

(cherry picked from commit fc5c6a5138474d01447e5d80b9e0a5478be6905b)