cpu_input_boost: Remove boost on framebuffer blank
There was a bug that caused the minfreq of the CPU to get stuck at policy->max while the screen was off and the CPU was awake. The bug is fixed now (see the change to fb_boost_fn()), but boosting on fb blank is too risky and not worth it, so remove it.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

(cherry picked from commit f39d8c2ca98e42a57225de84a9b485fec7f0c5e9)