msm: thermal: Add support to online the hotplugged cores
Add support in KTM to online the hotplugged cores when the
thermal condition is cleared.

In newer targets, KTM has the responsibility of
onlining the cores hotplugged due to thermal
condition. This feature is enabled by defining a device
tree property. If this property is not defined, KTM
does the legacy functionality where it doesn't
bring the core back online.

Change-Id: I3c75466774b309846df3176df6029b886bd59896
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
https://github.com/neobuddy89/shamu/commit/dd0e62a9a12a9bfb9596daad3c7cc1df75b03e10

 msm: thermal: Online only cores, which are hotplugged by KTM

On thermal hotplug clear event, online only cores which are offlined
by KTM. During a thermal hotplug event, KTM evaluates the thermal
hotplug status of the cpus and hotplugs or brings them up accordingly.
It impacts cases where some cores are not cold boot yet or those cores
are offlined due to non thermal reason.

Change-Id: Ib4073be13c28c444ff6d322432b4ed46e10b1f08
Signed-off-by: Manaf Meethalavalappu Pallikunhi <manafm@codeaurora.org>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
https://github.com/neobuddy89/shamu/commit/a02fd5620efedd1d2594521529049bf765d15394

 ARM: dts: 8084: Online cores that thermal driver offlines

* Let's trust thermal driver to do justice.

Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
https://github.com/neobuddy89/shamu/commit/2689ba938601a60c397edc4bc8220d87e651c0e3